Updated Public consult logic
Resolved bug of public consult not showing up on screen if one of the participants delete the slot when the consult Status is pending/confirme.

The intended consult size should remain unchanged unless the TA click confirm consult(reduce size of consult) in the event the size of the consult does not match the number of students involved.